ai_service/requirements.txt
flask
torch
transformers
pillow
numpy
faiss-cpu
pyvi
uwsgi